Split config module out of shared module
