+++ quinn-proto/src/transport_parameters.rs
@@ -17,9 +17,10 @@ use err_derive::Error;
 use crate::{
     cid_queue::CidQueue,
     coding::{BufExt, BufMutExt, UnexpectedEnd},
+    config::{EndpointConfig, ServerConfig, TransportConfig},
     crypto,
-    shared::{ConnectionId, EndpointConfig, ResetToken, ServerConfig},
-    Side, TransportConfig, TransportError, VarInt, MAX_CID_SIZE, RESET_TOKEN_SIZE,
+    shared::{ConnectionId, ResetToken},
+    Side, TransportError, VarInt, MAX_CID_SIZE, RESET_TOKEN_SIZE,
 };
 
 // Apply a given macro to a list of all the transport parameters having integer types, along with
